refactor(capability): lift provider capability matrix into neutral internal/capability
Move ServiceCapability, ProviderCapability, and the static capability
descriptor out of the build-tagged internal/gui package into a new neutral,
untagged package internal/capability, so the upcoming TUI (pure Go, no build
tags) and the GUI share one source of truth.

- ADD internal/capability/capability.go: ServiceCapability/ProviderCapability
  (field names and json tags byte-identical) plus the descriptor body as
  func All() []ProviderCapability.
- ADD internal/capability/capability_test.go: matrix invariants (Azure App
  Config unversioned + namespaced; force-delete/recovery-window AWS-secret
  only; per-version tags Azure Key Vault only; every service staged; restore
  on the soft-delete services; provider/service shape).
- MODIFY internal/gui/providers.go: keep the Wails binding surface stable via
  type aliases and Capabilities() delegating to capability.All(); everything
  else (detection/launch glue) untouched.
- MODIFY internal/gui/capabilities_internal_test.go: drop matrix-content
  assertions (now in the new package), keep the binding-delegation and
  ParamTypeOptions tests.
- deadcode-allow.txt: temporary waiver for capability.All (GUI-only caller in
  this step; Step 2's untagged TUI makes it reachable and removes the waiver).

Bindings unchanged: aliases resolve to the same JSON shape, so wailsjs models
need no regeneration (verified: gui DTO-contract test passes under -tags
production).
